+++ spec/truffle/interop/java_array_spec.rb
@@ -11,7 +11,7 @@ require_relative '../../ruby/spec_helper'
 describe "Truffle::Interop.java_array" do
 
   it "creates a Object[] from the values" do
-    Truffle::Debug.java_class_of(Truffle::Interop.deproxy(Truffle::Interop.java_array(1, 2, 3))).should == "Object[]"
+    Truffle::Interop.java_array(1, 2, 3).getClass.name.should == "java.lang.Object[]"
   end
 
   it "copies the correct values" do
@@ -22,9 +22,9 @@ describe "Truffle::Interop.java_array" do
   end
 
   it "will use Object[] because splat does not specialise for the type" do
-    Truffle::Debug.java_class_of(Truffle::Interop.deproxy(Truffle::Interop.java_array(1, 2, 3))).should == "Object[]"
-    Truffle::Debug.java_class_of(Truffle::Interop.deproxy(Truffle::Interop.java_array(1.1, 2.2, 3.3))).should == "Object[]"
-    Truffle::Debug.java_class_of(Truffle::Interop.deproxy(Truffle::Interop.java_array(:a, :b, :c))).should == "Object[]"
+    Truffle::Interop.java_array(1, 2, 3).getClass.name.should == "java.lang.Object[]"
+    Truffle::Interop.java_array(1.1, 2.2, 3.3).getClass.name.should == "java.lang.Object[]"
+    Truffle::Interop.java_array(:a, :b, :c).getClass.name.should == "java.lang.Object[]"
   end
 
   it "creates a copy" do