+++ CHANGELOG.md
@@ -66,6 +66,7 @@ Performance:
 
 Changes:
 
+* Remove `Truffle::Interop.deproxy` as it is unsafe and not useful (@eregon).
 
 # 22.3.0